Finished porting to 1.21.6-1.21.8

+++ common/src/main/java/einstein/subtle_effects/init/ModEntityTickers.java
@@ -32,10 +32,7 @@ import net.minecraft.world.entity.npc.AbstractVillager;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.entity.vehicle.AbstractMinecart;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.block.BaseFireBlock;
-import net.minecraft.world.level.block.Block;
-import net.minecraft.world.level.block.Blocks;
-import net.minecraft.world.level.block.BrushableBlock;
+import net.minecraft.world.level.block.*;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.Vec3;
 
@@ -97,7 +94,7 @@ public class ModEntityTickers {
             FallingBlockEntity fallingBlock = (FallingBlockEntity) entity;
 
             int startDistance = BLOCKS.fallingBlocks.whileFallingDustStartDistance.get();
-            float fallDistance = fallingBlock.fallDistance;
+            double fallDistance = fallingBlock.fallDistance;
             boolean isInWater = ((FallingBlockAccessor) fallingBlock).subtleEffects$isInWater() && BLOCKS.fallingBlocks.replaceDustWithBubblesUnderwater;
 
             if (fallDistance <= startDistance && !isInWater) {
